--- src/k_bios_term/Makefile
@@ -1,17 +1,0 @@
-
-
-# REQUIRED: Name of the standalone module (MUST BE THE SAME AS IT"S DIRECTORY NAME)
-MOD_NAME 	?= k_bios_term
-
-# REQUIRED: Names (NOT PATHS) of all .c files found in the src folder
-_SRCS 		?= term.c \
-			   term_sys_helpers.c
-
-# REQUIRED: Names (NOT PATHS) of all .s files found in the src folder
-_ASMS		?= 
-
-# REQUIRED: Names (NOT PATHS) of all .c files in the test folder
-_TEST_SRCS 	?= test.c
-
--include ../mod_stub.mk
-
