--- src/k_bios_term/test/test.c
@@ -1,4 +1,0 @@
-
-int test_term(void) {
-    return 1;
-}
